Store sitewide stats in a separate db
Currently sitewide stats are stored alongside user stats in a couchdb database.
When a day's user stats have been generated, the old day's database is removed,
deleting the older sitewide stats as well. Between then and the time today's
sitewide stats are generated they are unavailable to view on the site. To avoid
this issue store sitewide stats in a different db.