Improve modal typing in
The type for NiceModal.show are not convenient to use, requiring importing and using the NiceModalHOC type in conjunction with the props we are passing to the modal.
I couldn't find how to define a return type for the modal that we automatically be determined when passing the component to NiceModal.show.
The suggest method in the forums involves using async/await instead of a promise chain to set the return type manually. Not ideal still but simpler.
https://github.com/eBay/nice-modal-react/issues/127#issuecomment-1742452067