YIM: Tweak templates
Fix some issue with rendering the SVGs to canvas because of nested title elements.
Also fixes LB-1675 where links in the SVG are relative and point to non-existing pages on api.LB that should point to LB. Even better would be to detect beta-api.LB and test-api.LB and point to beta.LB and test.LB respectively.