Fix @crossdomain decorator
When handling OPTIONS requests, Flask can either generate an automatic response internally
or route the request to the view function for handling. The request is routed to the view
function if the view function has a provides_automatic_options attribute set to False or
if the methods list passed to route definition includes OPTIONS. OPTIONS requests
handled by Flask internally don't have CORS headers. Hence, API endpoints that need to
support CORS should ensure that the OPTIONS request is routed to the view function.

The @crossdomain decorator sets the provides_automatic_options attribute on the view
function which should have ensured this. However, due to the way decorators work the
attribute should have instead been set on the decorator function itself for Flask to
pick it up. However, we often list OPTIONS in the methods list of the route anyway
when applying @crossdomain. As a result, Flask would still route the request as expected.
The errors would only happen when the OPTIONS method was not mentioned in the route's
methods list.

@crossdomain decorator has now been fixed to set the attribute correctly and also add
OPTIONS to the methods list automatically. Further, a test has been added to ensure that
all API endpoints support CORS.