Accept potentially invalid dates
A user stumbled upon an issue with a partial date with a year and day but no month ("1993-??-01").
The browser's date parser does not like that at all and throws an error which breaks the page.
We need to check if the date is valid before parsing, in a few places in the code where we might receive such dates.