FreshReleases: show more intermediary dates
We have a system to calculate dates to show on the timeline that depends on the number of releases for each day.
If there is a long stretch of time without the required minimum releases, the timeline is suddently less useful and less clear.
These modifications ensure a tick is shown for at minimum every week.