Incremental Stats deployment fixes
Currently, the database for each stat is generated and the previous database for the
same stat is deleted daily. However, incremental stats generates messages only for
users with listens in incremental dumps, assuming stats for users without new listens
are already available on the LB server. To maintain this assumption, each stats database
must have a fixed name. Therefore, avoid sending the database name from the LB server to
Spark in stat generation requests. During stats generation, Spark determines whether it's
a full or incremental generation. For full stats, a new database name is created and sent;
for incremental stats, a database prefix is sent for the LB server to look up the existing
database and insert the stats accordingly.