Playlist page: Prevent autofocus on add-track component
This focus-on-load had a good intent, but it scrolls to the bottom of the page automatically which is annoying.
In some components it is desirable, but not for the playlist page, so I added an autofocus option that defaults to true (keeps it the same for other components)
There is a button at the top that takes you to the add-track component, so I made that trigger focus on the input component instead.