Art API: Fetch playlist image without set dimensions
We don't always have access to the right information from the front-end to make a decision about layout and dimensions of the cover art grid, if the user has not yet set it in the playlist.
This new endpoint allows fetching an image without specifying a layout, relying on the saved data if present in playlist metadata and defaulting safely otherwise.

Also gets rid of mentions of https://art.listenbrainz.org as that page no longer exists.