More granular incremental stats (#3164)
* Experiment more granular incremental stats filter

Currently, the entirety of incremental dumps is checked to filter entities whose
stats/popularity needs to be recomputed. This can be further optimized by storing
the latest created timestamp of the listens in incremental dumps when a stat is
run. The next time a stat is run, only listens with a higher created value (added
through newer incremental dumps) are considered for the filter.

Note that the incremental aggregate is still computed from all of incremental listens
dump, only the filter is made more granular.

* Update popularity providers

* fix SitewideStatsQueryProvider

* fix str-ing path where needed

* fix where clause

* fix where clause - 2

* debug

* debug - 2

* need to filter incremental aggregate too

* debug 3

* debug 4

* fix popularity generation

* fix function name in sitewide stats