Add disambiguation to track search results, Take 2
Did not like the formatting with disamb. in the title from previous commit.
There are no provisions for disambiguation in the Listen format, so implementing this directly in ListenCard does not make sense.
Instead, use the customization feature we already have in order to format the title and disamb. as desired.