Move created timestamp resolution reduction to postgres
created column of the listen table and listen_created column in the
listen_delete_metadata table are used in an anti join to filter deleted
listens in the spark cluster. The rounding off done by pyarrow seems to
be different from the rounding off done by spark/postgres due to which
some of the rows miss the join by 1 ms. To ensure consistency in rounding
and ensure that the created values match, do the resolution reduction for
both cases in postgres.