Fix get_intermediate_stats_df
When loading intermediate stats dataframe if start and end is not provided,
it reads the file as is without excluding year and month column used for
partitioning. This causes an error when trying to combining it with the
incremental listens dataframe which doesn't have the two fields. Hence,
fix the loading.