Update this_(week/month/year) time ranges
Consider this_week range and the example of week Feb 17 - Feb 24. If today
is Feb 20, currently this_week range resolves to Feb 17 - Feb 20 midnight.
Note that this range is used to filter listens using listened_at timestamp.

Imagine the two listens:

|          | listened_at         | created             |
|----------|---------------------|---------------------|
| Listen 1 | 2025-02-20 12:00:04 | 2025-02-20 12:01:55 |
| Listen 2 | 2025-02-18 09:00:00 | 2025-02-20 12:02:05 |

The second listen will be selected in the filter but not the first. The
incremental stats run records the max created during in bookkeeping and uses
this timestamp as the minimum created timestamp for the next incremental
stats run. As a result Listen 1 will not be included in any stats run. To
fix this, relax the this_week timestamp filters to the full week. this_week
range will now resolve to Feb 17 - Feb 24.

Similar logic applies to this_month and this_year.