Change dumps sequencing and misc test improvements
Currently, all dumps use the previous dumps' created timestamp as the start time
regardless of whether it is a full dump or an incremental dump. This is problematic
if we want to eliminate the use of full dumps and just rely on incremental dumps.
For example,

|                    | start               | end (created)       |
|--------------------|---------------------|---------------------|
| Incremental Dump 1 | 2025-01-30 12:00:01 | 2025-01-31 12:00:02 |
| Full Dump 1        | `LISTEN_MINIMUM_DATE` | 2025-02-01 04:00:05 |
| Incremental Dump 2 | 2025-02-01 04:00:05 | 2025-02-02 12:00:02 |

In this case if we were to just use incremental dumps, we would miss the listens
created between 2025-01-31 12:00:02 and 2025-02-01 04:00:05. To fix this issue,
update the start time of incremental dumps to use the end time of the last
incremental dump only. As currently it is not possible to identify from the
data_dump table whether a dump is incremental or full, update it to add dump_type
column.

Further, to ensure that full dumps don't contain duplicates with incremental dumps
add a where clause to full dumps query with a created check in addition to
listened_at.

Additionally, while diagnosing test failures made changes to the code to use datetime
aware objects everywhere. If dumps related tests fail silent and involve invocation of
click CliRunner, pass in catch_exceptions = False to debug.