Compact listen dumps in HDFS
Full dumps when imported in Spark are partitioned by listened_at's year
and month for storage in HDFS. Incremental dumps are imported everyday
and appended to a single incremental.parquet. Deleted listens are similarly
stored in deleted-listens.parquet and deleted-user-listen-history.parquet.
At run time, the full dumps are read and concatenated with incremental
dumps and the deleted listens are filtered out from the union. When a new
full dump is imported, it contains all the listens till that time and
all the deleted listens removed and the additional parquet files for
incremental and deleted listens are removed. This happens on a biweekly
timeline at the moment.

Full dumps are cumbersome to produce, hence we want to reduce our dependence
on them inside of ListenBrainz and Spark. After an initial full dump import
to seed the cluster, we intend to get rid of the biweekly full dump imports
and just rely on incremental dumps continuously. Hence, we need to rethink
some steps in how incremental listens are stored in the spark cluster and
how to implement deletions.

The solution I have come up with is replace the full dump import step with a
compaction step which reads all the partitioned base listens combines them
with incremental listens, removes the deleted listens and writes them back
to HDFS in the partitioned format. Everything else remains same.