Speedup statistics dumps
Using skip/limit for pagination of all documents in a couchdb is very slow. A
much better alternative is to paginate using start and end keys. For some reason,
in this case using startkey with skip/limit doesn't work but startkey_docid does.
As the key and document id in our case are the same, it works.