move back to spark_manage.py
if listenbrainz_spark/__main__.py is used as the file to execute, listenbrainz_spark is
added to the start of the sys.path. the biggest issue this causes is that listenbrainz_spark.hdfs
is resolved whenever the hdfs package from pypi is imported. further, its likely some imports in
other places break.