Mock DOMPurify dependency
DOMPurify is a browser-only library, and our testing harness is failing to import it when running in node.
Tried replacing it with isomorphic-dompurify but that didn't work either, so I suspect some deeper conflict with our JSDom setup for use with Jest.
Can't be bothered enough to dig tot he center of the world, so mocking the method we use as a passthrough function since we don't explicitly test its functionality