LB-1766: Fix Internal Server Error (500) when accessing listening history
In user.py, Timezone isnt being added to the datetime object (making it offset-naive) which is later sent to timescale_listenstore, where its getting compared with another datetime object with timezone i.e min_user_ts. This is leading to ambigous behaviour for users with less amount of listens.

Add UTC timezone to both from_ts and to_ts (similar to how we do it in api.py ) .