Disable playlist export buttons instead of hiding them
I think for now it is nicer to keep the same UI when not logged in but disable the buttons if missing appropriate credentials.

Maybe we will revisit this when we have more music services in there so as not to overload the menu with a bunch of options the user can't click on.