Render placeholder for older thank you event
We currently show a preview of the original recommendation using the events currently shown on the page, basically rendering it again under the thank you event.
But if the event is older than the current events shown on the page, it does not render anything, which is  confusing for users.
Show a placeholder instead of nothing, until we have a better way to fetch the event in question.