Catch image load errors when using Vibrant
Have been seeing some errors after introducing Vibrant to get color palettes.
The issue is that if image loading fails, an error is thrown by the library and not caught in our code, resulting in an error page.
Catching the error and logging to console ensures we don;t end up with an error page, although i will need to investigate the root cause of the issue (CORS errors it looks like, sob).