Use HTMX for pin widget
This moves all the logic and templating from some dodgy javascript written in an html file to Flask and Jinja templates.
Simplifies things a lot, and reduces how much JS we need to load or copy in order to format things properly.
The templating is also much better and safer this way.