Widgets: Go back to timer polling instead of websockets
Can't really see a lot of advantages with loading a JS library for this (considering we use socket.io, which is not supported by HTMX), when we can just do some regular polling very,m very easily.

Kept the other implementation tests commented out for if we wan tot revisit this after more research.