Fix event_type enum serialization
json serialization of enum string values fails, forcing us to manually serialize the value.
Didn't do it properly in a recent PR, so looked into it closer to avoid future copy-paste issues or just not knowing.

We can use the class StrEnum (added in python 3.11) and that will serialize just fine...