Widgets: Pin: fallback to previous pin and remove HTMX
Adds a fallback to the last pin if there is no current pin and adds conditional styling for the card (currently pinned = blue BG).

The HTMX back-and-forth was not adding anything other than a javascript library loaded for nothing, since we can simply render the right template.
There is no need for auto-refresh or websockets or anything, so let's keep it simple for that widget and save some bandwidth.

+ a couple other CSS tweaks suggested in the PR review.