Add sample data dumps
Preliminary version of sample data dumps to help develop ListenBrainz locally.
These dumps include metadata and popularity data used on entity pages and in
showing enriched metadata on ListenBrainz web pages.

I intend to include listens using this metadata and similarity in a future
iteration of the sample dumps. The incremental dumps might not be good enough
as the entity mbids of listens in an incremental dump might not be present
in the small subset of data that is included in sample dumps.