Show latest imported listens' timestamp instead of import progress
Import progress would not work well because we only fetch the total listen count
when the listen import is first enqueued. On incremental imports, the value is
always set to 0 by the frontend as a default and the import progress is invalidly
set to 0%. We could fetch the total count everytime from last.fm when the import
status is refreshed, but I feel the last imported listen which we store in LB
itself.