Improve incremental dump error reporting
Incremental dumps report an error when any dump id is missing or could not be
imported since the last full dump id. This used to work fine because always
had the latest full dump available and only imported incremental dumps since
then. However, now spark doesn't regularly import full dumps and there are
gaps in the incremental dumps id as a result. Therefore, to avoid false errors
where the gap is expected, check if the gap is because of a full dump id and
ignore it.