Rename MusicServicePermissionsBox to LFMMusicServicePermissions
For the moment, this component is used solely for use with Last.FM and Libre.FM, so renaming this component to make it a  bit clearer.
We already have a generic 'ServicePermissionButton' and I was worried the too-generic name would create confusion.

We might revisit this if we end up having tho reuse this specialized component for another service.