Fix "Invalid video id" in YoutubePlayer
While using a debugger to trace the issue, I found the following line inside
the Youtube's iFramePlayer code triggered the issue:
`if((a=String(Y(a,"videoId")))&&(a.length!==11||!a.match(/^[a-zA-Z0-9\-_]+$/)))throw Error("Invalid video id");`

I am guessing that code is basically stringifying the input videoId, String(null)
evaluates to "null" and String(undefined) to "undefined". Empty string avoids the issue,
hence using that as default videoId.