Reinstate "import listens" page
This page was removed in #3301 as we moved the last feature out of that page.
However we have an ongoing GSOC project for importing backup files, which will need a place to live. This is the page.
Modified the mostly empty page to make it a  better placeholder and clearly point ot other pages where needed.
Since the removal has not been deployed yet, rather than re-adding it later we are opting for re-adding it now with a placeholder so the page never officially went away.