Mock react-markdown
After updating the dependency react-markdown to v10 in commit 7918a4d , the tests are failing because react-markdown does not export a compiled CommonJS file anymore and only exports an ESM module, which Jest does not like and throws an error: "SyntaxError: Unexpected token 'export'"

See the issue https://github.com/remarkjs/react-markdown/issues/635 for context.

Updating the Jest configuration to point to a custom mock file instead, as we don't directly test the output fo react-markdown.