Fix potentially broken page when datasource missing
As we are adding new music datasources, in the process you may find yourself with one datasource information in the database that is does not exist ijn the version currently deployed (for example, test a new music service on test then going to the prod website).
Since the datasource is missing, these lines break the BrainzPlayer settings page completely.
Making it safe access to drilled property.