Remove show prop from music players
This 'show' prop was a hacky way of stoppingmusic sources and hiding their content, and was causing some issues with re-rendering lifecycle, preventing correct playback.

Now that we can use Jotai to synchronize state instead, removing this prop in favor of a new  'stop' function on datasources, which can be called by the parent before playing the next track.

Creates a  currentDataSourceNameAtom so each datasource can determine if it is the current source

+ a couple of other small refactor/simplification