Fix SpotifyPlayer connection and retry mechanisms
Ensure thye Spotify player has the time to set itself up properly beofre trying to play, toherwise we get interference with itself trying to reconnect/recreate the player in a loop.
Also store deviceID as a class property instead of react state, since it is not used for rendering and it introduces a re-rendering delay.