Cache get_listens API endpoint
Implement caching for get_listens endpoint. The cache keys for the API call consider
various query params and all the cached responses are stored in a per user set in the
cache as well. When the cache has to be invalidated for listen insertion, deletion or
manual linking, all the listen cache keys are retrieved from the set and are invalidated.
The set having all cache keys is set to expire at twice the duration of listen cache
expiration to ensure it is available when needed for invalidation.