Album grids: force crop each image to its defined width and height
Canvg (library we use for rendering SVG to image) has problems with cropping images to size, causing some overlapping for non-square images in the grids despite the preserveAspectRatio "slice" mode.

Generating a clipPath element for each image and using that as clip-path resolves this issue. It's cumbersome, but it ensures our squares are square.