Handle unsupported service properly in background tasks
Currently, when background listens importer comes across an import task from a service
it cannot recognize, it deletes the entry in the background task table but not in the
listens importer table. As the listens importer entry is left with status waiting, this
creates issues when creating a new import. Also, the user is shown a waiting import entry
for a file that will never be processed as the background task has already been deleted.
Update the tasks metadata with an error in this case.

Rest of the changes are mostly rearranging the code for ease of use in the test.